dclient/requirements.txt
discord-py
websockets
pyyaml
requests
